09/09_02_develop_script/app/requirements.txt
mysql-connector-python==8.0.29  # GPL
